+++ srv/src/main/java/sap/capire/emoji_plugin/Application.java
@@ -1,4 +1,4 @@
-package sap.capire.emoji_template;
+package sap.capire.emoji_plugin;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
