+++ srv/src/main/resources/META-INF/services/com.sap.cds.services.runtime.CdsRuntimeConfiguration
@@ -1,1 +1,1 @@
-sap.capire.emoji_template.EmojiHandlerRuntimeConfiguration
+sap.capire.emoji_plugin.EmojiHandlerRuntimeConfiguration
